reference/pyproject.toml
[project]
name = "grounding-reference"
version = "0.1.0"
description = "Reference implementation of the grounding pattern (triage → research → writer)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.40",
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]
